printf almost done

+++ projects/ft_printf/headers/ft_eval_utils.h
@@ -6,11 +6,11 @@
 
 typedef	struct s_number
 {
-	unsigned int	u_num;
-	int 			sign;
+	unsigned long long	u_num;
+	int 				sign;
 }	t_number;
 
-void	ft_process_number_(t_number *n, long int num);
+void	ft_process_number_(t_number *n, long long num);
 int		ft_add_precision_(t_conv *conv);
 int 	ft_set_prefix_(t_conv *conv, t_number n);
 int 	ft_set_shift_(t_conv *conv);
